Add webhook server for GitHub App integration
Extract shared logic (log parsing, table generation, comment posting)
into src/core.ts. Add src/server.ts that listens for workflow_run.completed
webhooks so users just install the GitHub App — no secrets or workflow
config needed. The action entry point (src/index.ts) still works too.

Users configure per-repo via .github/cpp-warning-notifier.json:
  { "job_regex": "...", "step_regex": "...",
    "row_headers": [...], "column_header": "..." }

Server env vars: APP_ID, APP_PRIVATE_KEY, WEBHOOK_SECRET, PORT.
