+++ rollup.config.ts
@@ -2,7 +2,9 @@ import commonjs from "@rollup/plugin-commonjs";
 import nodeResolve from "@rollup/plugin-node-resolve";
 import typescript from "@rollup/plugin-typescript";
 
-const config = {
+const plugins = [typescript(), nodeResolve({ preferBuiltins: true }), commonjs()];
+
+const action = {
   input: "src/index.ts",
   output: {
     esModule: true,
@@ -10,7 +12,18 @@ const config = {
     format: "es",
     sourcemap: true,
   },
-  plugins: [typescript(), nodeResolve({ preferBuiltins: true }), commonjs()],
+  plugins,
 };
 
-export default config;
+const server = {
+  input: "src/server.ts",
+  output: {
+    esModule: true,
+    file: "dist/server.js",
+    format: "es",
+    sourcemap: true,
+  },
+  plugins,
+};
+
+export default [action, server];
